requirements.txt
fonttools==4.42.1
matplotlib==3.7.2
mmcv==2.0.1
mmengine==0.8.4
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
torch==2.0.1
torchvision==0.15.2
gdown==4.7.1
h5py==3.8.0
mediapipe==0.9.1.0
pandas==1.5.3
pathlib==1.0.1
scikit-image==0.20.0
tqdm==4.64.1
wandb==0.15.7
